implement basic tokenization
this tokenization is not optimized against
splits lengths and their frequency.

This kind of tokenization is implemented
for tokenizers with direct tokenization procedure
 like farasa and disjoint.

For morphological, this does not apply
as the tokenizer is already trained.

For random, this also does not apply due to randomization.

For words and characters,
they are already implemented with such kind of tokenization.